+++ x/evm/keeper/state_transition.go
@@ -65,6 +65,7 @@ func (k *Keeper) NewEVM(
 		Difficulty:  cfg.Difficulty,
 		BaseFee:     cfg.BaseFee,
 		Random:      cfg.Random,
+		BlobBaseFee: cfg.BlobBaseFee,
 	}
 	if cfg.BlockOverrides != nil {
 		cfg.BlockOverrides.Apply(&blockCtx)